Move `Grants` to `NylasClient` and custom authentication to `Auth` (#324)
* Move custom auth out of grants to auth

* move grants entry point out of auth to nylasclient

* Update CHANGELOG.md
